Entrega final de escritorio
Descargar y probar para que se vean errores y me digan

+++ PacketWorldEscritorio/src/packetworldescritorio/FXMLSucursalController.java
@@ -16,8 +16,10 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.AnchorPane;
+import packetworldescritorio.modelo.dao.EnviosDAO;
 import packetworldescritorio.pojo.Mensaje;
 import packetworldescritorio.modelo.dao.SucursalDAO;
+import packetworldescritorio.pojo.Envio;
 import packetworldescritorio.pojo.Sucursal;
 import packetworldescritorio.utilidades.Alertas;
 import static packetworldescritorio.utilidades.Alertas.mostrarAlertaConfirmacion;
@@ -58,7 +60,7 @@ public class FXMLSucursalController implements Initializable {
         configurarTabla();
         cargarInformacion();
     }
-    
+
     private void configurarTabla() {
         colCodigo.setCellValueFactory(new PropertyValueFactory("codigo"));
         colNombreCorto.setCellValueFactory(new PropertyValueFactory("nombreCorto"));
@@ -72,7 +74,7 @@ public class FXMLSucursalController implements Initializable {
         colCiudad.setCellValueFactory(new PropertyValueFactory("ciudad"));
         colEstado.setCellValueFactory(new PropertyValueFactory("estado"));
     }
-    
+
     private void cargarInformacion() {
         sucursales = FXCollections.observableArrayList();
         List<Sucursal> WSList = SucursalDAO.obtenerSucursales();
@@ -110,9 +112,32 @@ public class FXMLSucursalController implements Initializable {
     private void btnEliminar(ActionEvent event) {
         Sucursal sucursal = tablaSucursales.getSelectionModel().getSelectedItem();
         if (sucursal != null) {
-            eliminarSucursal(sucursal);
+            if (sucursal.getEstatus().equals("Inactiva")) {
+                boolean confirmado = Alertas.mostrarAlertaConfirmacion("Confirmar eliminación", "¿Esta seguro de querer eliminar esta sucursal?\nYa se encuentra inactiva, por lo que eliminarla hara que desaparezca de los registros.");
+                if (confirmado) {
+                    eliminarSucursal(sucursal);
+                }
+            } else {
+                desactivarSucursal(sucursal);
+            }
         } else {
             Alertas.mostrarAlertaSimple("Seleccionar sucursal.", "Para eliminar debes seleccionar un sucursal de la tabla.", Alert.AlertType.WARNING);
+        }
+    }
+
+    private void desactivarSucursal(Sucursal sucursal) {
+        Mensaje msj = new Mensaje();
+        boolean confirmado = mostrarAlertaConfirmacion("Confirmar eliminación",
+                "¿Está seguro de que desea dar de baja esta sucursal?");
+        if (confirmado) {
+
+            msj = SucursalDAO.darBajaSucursal(sucursal.getIdSucursal());
+            if (!msj.isError()) {
+                Alertas.mostrarAlertaSimple("Sucursal eliminada", "La sucursal ha sido eliminado correctamente.", Alert.AlertType.INFORMATION);
+                cargarInformacion();
+            } else {
+                Alertas.mostrarAlertaSimple("Error al eliminar.", "No se pudo eliminar la sucursal, intente nuevamente.", Alert.AlertType.WARNING);
+            }
         }
     }
 
@@ -121,10 +146,17 @@ public class FXMLSucursalController implements Initializable {
         boolean confirmado = mostrarAlertaConfirmacion("Confirmar eliminación",
                 "¿Está seguro de que desea eliminar esta sucursal?");
         if (confirmado) {
-            confirmado = mostrarAlertaConfirmacion("Confirmar eliminación",
-                    "¿Realmente está seguro de que desea eliminar esta sucursal?");
-            if (confirmado) {
-                msj = SucursalDAO.darBajaSucursal(sucursal.getIdSucursal());
+            List<Envio> envios = EnviosDAO.obtenerEnvios();
+            boolean tieneEnvios = false;
+            for (Envio envio : envios) {
+                if (envio.getIdSucursalOrigen() == sucursal.getIdSucursal()) {
+                    tieneEnvios = true;
+                }
+            }
+            if (tieneEnvios) {
+                Alertas.mostrarAlertaSimple("Problemas de eliminación", "La sucursal no se puede eliminar debido a que tiene envios asignados a ella. Borre los envios que corresponden a esta sucursal para poder eliminarla", Alert.AlertType.ERROR);
+            } else {
+                msj = SucursalDAO.eliminarSucursal(sucursal.getIdSucursal());
                 if (!msj.isError()) {
                     Alertas.mostrarAlertaSimple("Sucursal eliminada", "La sucursal ha sido eliminado correctamente.", Alert.AlertType.INFORMATION);
                     cargarInformacion();
@@ -132,8 +164,6 @@ public class FXMLSucursalController implements Initializable {
                     Alertas.mostrarAlertaSimple("Error al eliminar.", "No se pudo eliminar la sucursal, intente nuevamente.", Alert.AlertType.WARNING);
                 }
             }
-        } else {
-            Alertas.mostrarAlertaSimple("Error al eliminar.", "No se pudo eliminar el sucursal, intente nuevamente.", Alert.AlertType.WARNING);
         }
     }
